requirements.txt
absl-py==1.0.0
attrs==22.1.0
blis==0.7.8
cachetools==5.1.0
catalogue==1.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
cycler==0.11.0
cymem==2.0.6
fonttools==4.33.3
google-auth==2.6.6
google-auth-oauthlib==0.4.6
grpcio==1.46.1
idna==3.3
importlib-metadata==4.11.3
joblib==1.1.0
jsonschema==4.16.0
kiwisolver==1.4.2
Markdown==3.3.7
matplotlib==3.5.2
murmurhash==1.0.8
numpy==1.22.3
oauthlib==3.2.0
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.2
Pillow==9.1.1
plac==1.1.3
preshed==3.0.7
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyFFTW==0.13.0
pyparsing==3.0.9
pyproject-toml==0.0.10
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.1.0
scipy==1.8.0
six==1.16.0
srsly==1.0.5
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
thinc==7.4.5
threadpoolctl==3.1.0
toml==0.10.2
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.9
wasabi==0.10.1
Werkzeug==2.1.2
zipp==3.8.0
